update section in menu official and activity
