+++ apps-client/mobile/app/(tabs)/(official)/official.tsx
@@ -1,10 +1,7 @@
 import React, { useCallback, useMemo, useState } from 'react'
 import { Pressable, PressableProps, ScrollView, View } from 'react-native'
 import Animated, { useSharedValue, withTiming } from 'react-native-reanimated'
-import { useSafeAreaInsets } from 'react-native-safe-area-context'
 
-import { BottomSheetModal, BottomSheetView } from '@pple-today/ui/bottom-sheet/index'
-import { Button } from '@pple-today/ui/button'
 import { Icon } from '@pple-today/ui/icon'
 import { cn } from '@pple-today/ui/lib/utils'
 import {
@@ -17,30 +14,14 @@ import {
 } from '@pple-today/ui/select'
 import { Slide, SlideIndicators, SlideItem, SlideScrollView } from '@pple-today/ui/slide'
 import { Text } from '@pple-today/ui/text'
-import { H1, H2, H3 } from '@pple-today/ui/typography'
+import { H1, H2 } from '@pple-today/ui/typography'
 import { useQueryClient } from '@tanstack/react-query'
 import { Image } from 'expo-image'
-import * as Linking from 'expo-linking'
 import { useRouter } from 'expo-router'
-import * as WebBrowser from 'expo-web-browser'
-import {
-  ArrowRightIcon,
-  CircleChevronRightIcon,
-  GlobeIcon,
-  InfoIcon,
-  LandmarkIcon,
-  MailIcon,
-  MegaphoneIcon,
-  PanelsTopLeftIcon,
-  PhoneIcon,
-  VoteIcon,
-} from 'lucide-react-native'
+import { CircleChevronRightIcon, VoteIcon } from 'lucide-react-native'
 
-import ContactMail from '@app/assets/contact-mail.svg'
 import Personal from '@app/assets/personal.svg'
 import PPLEIcon from '@app/assets/pple-icon.svg'
-import { UserAddressInfoSection } from '@app/components/address-info'
-import { AnnouncementCard, AnnouncementCardSkeleton } from '@app/components/announcement'
 import { ElectionCard } from '@app/components/election/election-card'
 import { RefreshControl } from '@app/components/refresh-control'
 import { SafeAreaLayout } from '@app/components/safe-area-layout'
@@ -50,15 +31,16 @@ import { getRoleName } from '@app/utils/get-role-name'
 
 import { useBottomTabOnPress } from '../_layout'
 
+const ALL_ROLES_OPTION: Option = { value: '', label: 'ทั้งหมด' }
+
 export default function OfficialPage() {
   const queryClient = useQueryClient()
+  const [selectedRole, setSelectedRole] = useState<Option | undefined>(ALL_ROLES_OPTION)
+  const profileQuery = reactQueryClient.useQuery('/profile/me', {})
   const onRefresh = useCallback(async () => {
     await Promise.all([
       queryClient.invalidateQueries({
         queryKey: reactQueryClient.getQueryKey('/elections'),
-      }),
-      queryClient.resetQueries({
-        queryKey: reactQueryClient.getQueryKey('/announcements'),
       }),
       queryClient.resetQueries({
         queryKey: reactQueryClient.getQueryKey('/mini-app'),
@@ -79,25 +61,32 @@ export default function OfficialPage() {
         refreshControl={<RefreshControl onRefresh={onRefresh} />}
       >
         <View className="flex flex-col p-4 bg-base-bg-white">
-          <View className="flex flex-row gap-2 items-center">
-            <Icon
-              icon={LandmarkIcon}
-              size={32}
-              strokeWidth={2}
-              className="text-base-primary-default"
-            />
-            <H1 className="text-3xl font-heading-semibold text-base-primary-default">ทางการ</H1>
+          <View className="flex flex-row gap-2 items-center justify-between">
+            <View className="flex flex-row gap-2 items-center">
+              <PPLEIcon width={35} height={30} />
+              <H1 className="text-3xl font-heading-semibold text-base-primary-default">แอป</H1>
+            </View>
+            {profileQuery.data && profileQuery.data.roles.length > 0 && (
+              <Select value={selectedRole} onValueChange={setSelectedRole}>
+                <SelectTrigger className="min-w-[120px]">
+                  <SelectValue placeholder="ทั้งหมด" />
+                </SelectTrigger>
+                <SelectContent>
+                  <SelectItem label="ทั้งหมด" value="" />
+                  {profileQuery.data.roles.map((role) => (
+                    <SelectItem key={role} label={getRoleName([role])} value={role} />
+                  ))}
+                </SelectContent>
+              </Select>
+            )}
           </View>
           <Text className="font-heading-regular text-base-text-medium">
-            ข้อมูลข่าวสารจากพรรคประชาชน
+            แอปพลิเคชันจากพรรคประชาชน
           </Text>
         </View>
-        <UserAddressInfoSection className="pb-4 bg-base-bg-white" />
         <View className="gap-3 py-4">
           <ElectionSection />
-          <AnnouncementSection />
-          <InformationSection />
-          <MiniAppSection />
+          <MiniAppSection selectedRole={selectedRole} />
         </View>
       </ScrollView>
     </SafeAreaLayout>
@@ -142,244 +131,8 @@ const ElectionSection = () => {
   )
 }
 
-const AnnouncementSection = () => {
+const MiniAppSection = ({ selectedRole }: { selectedRole: Option | undefined }) => {
   const router = useRouter()
-  const announcementsQuery = reactQueryClient.useQuery('/announcements', {
-    query: { limit: 3 },
-  })
-
-  const data = announcementsQuery.data?.announcements
-
-  const AnnouncementPreviewList = () => {
-    if (announcementsQuery.isLoading || !data) {
-      return (
-        <View className="my-3 gap-3">
-          <AnnouncementCardSkeleton className="w-full" />
-          <AnnouncementCardSkeleton className="w-full" />
-          <AnnouncementCardSkeleton className="w-full" />
-        </View>
-      )
-    }
-
-    return (
-      <View className="my-3 gap-3">
-        {data.map((item) => (
-          <AnnouncementCard
-            className="w-full"
-            key={item.id}
-            onPress={() => router.navigate(`/announcement/${item.id}`)}
-            id={item.id}
-            feedId={item.id}
-            title={item.title}
-            date={item.publishedAt.toString()}
-            type={item.type}
-          />
-        ))}
-        {data.length === 0 && (
-          <View className="flex flex-col items-center justify-center">
-            <Text className="text-base-text-medium font-heading-semibold">ยังไม่มีประกาศ</Text>
-          </View>
-        )}
-      </View>
-    )
-  }
-
-  return (
-    <View className="px-4">
-      <View className="flex flex-row justify-between items-center">
-        <View className="flex flex-row gap-2 items-center">
-          <Icon icon={MegaphoneIcon} size={32} className="text-base-primary-default" />
-          <H2 className="text-2xl font-heading-semibold text-base-text-high">ประกาศ</H2>
-        </View>
-        <View className="min-h-10">
-          {data && data.length > 0 && (
-            <Button variant="ghost" onPress={() => router.navigate('/announcement')}>
-              <Text>ดูเพิ่มเติม</Text>
-              <Icon icon={ArrowRightIcon} strokeWidth={2} />
-            </Button>
-          )}
-        </View>
-      </View>
-      <AnnouncementPreviewList />
-    </View>
-  )
-}
-
-const InformationSection = () => {
-  const bottomSheetModalRef = React.useRef<BottomSheetModal>(null)
-  const onOpen = () => {
-    bottomSheetModalRef.current?.present()
-  }
-
-  const insets = useSafeAreaInsets()
-
-  return (
-    <View className="px-4">
-      <View className="flex flex-row gap-2 items-center">
-        <View className="w-8 h-8 flex items-center justify-center">
-          <Icon icon={InfoIcon} size={32} className="text-base-primary-default" />
-        </View>
-        <H2 className="text-2xl font-heading-semibold text-base-text-high">ข้อมูลพรรคประชาชน</H2>
-      </View>
-      <View className="mt-4 gap-y-4">
-        <View className="flex flex-row gap-x-[12.5px]">
-          <InfoItem
-            onPress={() => WebBrowser.openBrowserAsync('https://peoplesparty.or.th/person/')}
-          >
-            <View className="flex justify-start flex-col flex-wrap">
-              <View className="flex flex-col mb-3 h-8 w-8 bg-base-secondary-default rounded-lg items-center justify-center">
-                <Personal width={24} height={24} color="white" />
-              </View>
-              <Text className="text-base font-heading-semibold w-full">บุคลากรของพรรค</Text>
-            </View>
-            <View className="flex-1 justify-end items-end">
-              <Icon
-                icon={CircleChevronRightIcon}
-                size={28}
-                strokeWidth={1}
-                className="text-foreground"
-              />
-            </View>
-          </InfoItem>
-          <InfoItem
-            onPress={() => WebBrowser.openBrowserAsync('https://peoplesparty.or.th/about/')}
-          >
-            <View className="flex justify-start flex-col flex-wrap">
-              <View className="flex flex-col mb-3 h-8 w-8 bg-base-primary-default rounded-lg items-center justify-center">
-                <PPLEIcon width={20} height={16} color="white" />
-              </View>
-              <Text className="text-base font-heading-semibold w-full">เกี่ยวกับพรรคประชาชน</Text>
-            </View>
-            <View className="flex-1 justify-end items-end">
-              <Icon
-                icon={CircleChevronRightIcon}
-                size={28}
-                strokeWidth={1}
-                className="text-foreground"
-              />
-            </View>
-          </InfoItem>
-        </View>
-        <View className="flex flex-row gap-x-[12.5px]">
-          <InfoItem onPress={onOpen}>
-            <View className="flex justify-start flex-col flex-wrap">
-              <View className="flex flex-col mb-3 h-8 w-8 bg-violet-500 rounded-lg items-center justify-center">
-                <ContactMail width={24} height={24} color="white" />
-              </View>
-              <Text className="text-base font-heading-semibold w-full">ช่องทางการติดต่อ</Text>
-            </View>
-            <View className="flex-1 justify-end items-end">
-              <Icon
-                icon={CircleChevronRightIcon}
-                size={28}
-                strokeWidth={1}
-                className="text-foreground"
-              />
-            </View>
-            <BottomSheetModal ref={bottomSheetModalRef} bottomInset={insets.bottom}>
-              <BottomSheetView>
-                <View className="p-4 pb-0">
-                  <H3>ช่องทางการติดต่อ</H3>
-                </View>
-                <View className="p-4">
-                  <View className="w-full rounded-lg border border-base-outline-default p-4 bg-base-bg-default">
-                    <View className="gap-4">
-                      <View className="gap-2">
-                        <View className="flex flex-row justify-start items-center gap-1">
-                          <Icon
-                            icon={LandmarkIcon}
-                            size={16}
-                            strokeWidth={1}
-                            className="text-base-primary-default"
-                          />
-                          <Text className="text-base-text-high font-body-medium">สำนักงานใหญ่</Text>
-                        </View>
-                        <View className="gap-0.5">
-                          <Text className="font-body-light">เลขที่ 167 อาคารอนาคตใหม่ ชั้น 4</Text>
-                          <Text className="font-body-light">
-                            รามคำแหง 42 แขวงหัวหมาก เขต บางกะปิ
-                          </Text>
-                          <Text className="font-body-light">กรุงเทพมหานคร 10240</Text>
-                        </View>
-                      </View>
-                      <View className="gap-2">
-                        <View className="flex flex-row justify-start items-center gap-1">
-                          <Icon
-                            icon={PhoneIcon}
-                            size={16}
-                            strokeWidth={1}
-                            className="text-base-primary-default"
-                          />
-                          <Text className="text-base-text-high font-body-medium">เบอร์โทร</Text>
-                        </View>
-                        <Text className="font-body-light items-baseline">
-                          <Text
-                            onPress={() => Linking.openURL('tel:028215874')}
-                            className="underline font-body-light"
-                          >
-                            02-821-5874
-                          </Text>{' '}
-                          (จันทร์-ศุกร์ 10:00-18:00 น.)
-                        </Text>
-                      </View>
-                      <View className="gap-2">
-                        <View className="flex flex-row justify-start items-center gap-1">
-                          <Icon
-                            icon={MailIcon}
-                            size={16}
-                            strokeWidth={1}
-                            className="text-base-primary-default"
-                          />
-                          <Text className="text-base-text-high font-body-medium">อีเมล</Text>
-                        </View>
-                        <View className="gap-0.5 mb-10">
-                          <Text
-                            className="font-body-light underline"
-                            onPress={() =>
-                              Linking.openURL('mailto:office@peoplespartythailand.org')
-                            }
-                          >
-                            office@peoplespartythailand.org
-                          </Text>
-                          <Text className="font-body-light">PeoplesPartyThailand</Text>
-                          <Text className="font-body-light">@PPLEThailand</Text>
-                          <Text className="font-body-light">พรรคประชาชน - People&apos;s Party</Text>
-                        </View>
-                      </View>
-                    </View>
-                  </View>
-                </View>
-              </BottomSheetView>
-            </BottomSheetModal>
-          </InfoItem>
-          <InfoItem onPress={() => WebBrowser.openBrowserAsync('https://peoplesparty.or.th/')}>
-            <View className="flex justify-start flex-col flex-wrap">
-              <View className="flex flex-col mb-3 h-8 w-8 bg-blue-500 rounded-lg items-center justify-center">
-                <Icon icon={GlobeIcon} size={22} color="white" />
-              </View>
-              <Text className="text-base font-heading-semibold w-full">เว็บไซต์ทางการ</Text>
-            </View>
-            <View className="flex-1 justify-end items-end">
-              <Icon
-                icon={CircleChevronRightIcon}
-                size={28}
-                strokeWidth={1}
-                className="text-foreground"
-              />
-            </View>
-          </InfoItem>
-        </View>
-      </View>
-    </View>
-  )
-}
-
-const ALL_ROLES_OPTION: Option = { value: '', label: 'ทั้งหมด' }
-
-const MiniAppSection = () => {
-  const router = useRouter()
-  const [selectedRole, setSelectedRole] = useState<Option | undefined>(ALL_ROLES_OPTION)
-  const profileQuery = reactQueryClient.useQuery('/profile/me', {})
   const { data: miniAppData } = reactQueryClient.useQuery('/mini-app', {
     query: selectedRole?.value ? { role: selectedRole.value } : {},
   })
@@ -403,27 +156,6 @@ const MiniAppSection = () => {
 
   return (
     <View className="px-4">
-      <View className="flex flex-row justify-between items-center">
-        <View className="flex flex-row gap-2 items-center">
-          <View className="w-8 h-8 flex items-center justify-center">
-            <Icon icon={PanelsTopLeftIcon} size={32} className="text-base-primary-default" />
-          </View>
-          <H2 className="text-2xl font-heading-semibold text-base-text-high">มินิแอปฯ</H2>
-        </View>
-        {profileQuery.data && profileQuery.data.roles.length > 0 && (
-          <Select value={selectedRole} onValueChange={setSelectedRole}>
-            <SelectTrigger className="min-w-[120px]">
-              <SelectValue placeholder="ทั้งหมด" />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem label="ทั้งหมด" value="" />
-              {profileQuery.data.roles.map((role) => (
-                <SelectItem key={role} label={getRoleName([role])} value={role} />
-              ))}
-            </SelectContent>
-          </Select>
-        )}
-      </View>
       <View className="flex flex-col gap-4 mt-4 items-center">
         {miniAppGroupByTwo.map((group, index) => (
           <View key={index} className="flex flex-1 flex-row gap-4">
